+++ apps/client-ts/src/components/Configuration/Catalog/CopySnippet.tsx
@@ -25,8 +25,8 @@ export const CopySnippet = () => {
 
     const handleCopyLeft = () => {
         navigator.clipboard.writeText(
-            `<PanoraDynamicCatalog
-                category={"crm"}
+            `<PanoraDynamicCatalogCard
+                category={ConnectorCategory.Crm}
                 projectId={"c9a1b1f8-466d-442d-a95e-11cdd00baf49"}
                 returnUrl={"https://acme.inc"}
                 linkedUserId={"b860d6c1-28f9-485c-86cd-fb09e60f10a2"}
@@ -48,7 +48,7 @@ export const CopySnippet = () => {
         navigator.clipboard.writeText(
             `<PanoraProviderCard
             name={"hubspot"}
-            category={"crm"}
+            category={ConnectorCategory.Crm}
             projectId={"c9a1b1f8-466d-442d-a95e-11cdd00baf49"}
             returnUrl={"https://acme.inc"}
             linkedUserId={"b860d6c1-28f9-485c-86cd-fb09e60f10a2"} 
@@ -121,11 +121,11 @@ export const CopySnippet = () => {
                     <span className="flex flex-col gap-2">
                         <span className="flex-1">
                         <span>
-                            {`<PanoraDynamicCatalog>`}
+                            {`<PanoraDynamicCatalogCard>`}
                         </span>
                         </span>
                         <span className="pl-4">
-                            <span className="text-green-800">category</span>{`=crm`}
+                            <span className="text-green-800">category</span>{`={ConnectorCategory.Crm}`}
                         </span>
                         <span className="pl-4">
                             <span className="text-cyan-600">projectId</span>{`={'c9a1b1f8-466d-442d-a95e-11cdd00baf49'}`}
@@ -138,7 +138,7 @@ export const CopySnippet = () => {
                         </span> 
                         <span className="flex-1">
                         <span>
-                            {`</PanoraDynamicCatalog>`}
+                            {`/>`}
                         </span>
                         </span>
                     </span>
@@ -185,7 +185,7 @@ export const CopySnippet = () => {
                       <span className="text-lime-500">name</span>{`={'hubspot'}`}
                     </span>
                     <span className="pl-4">
-                        <span className="text-fuchsia-200">category</span>{`={'crm'}`}
+                        <span className="text-fuchsia-200">category</span>{`={ConnectorCategory.Crm}`}
                     </span>
                     <span className="pl-4">
                         <span className="text-cyan-600">projectId</span>{`={'c9a1b1f8-466d-442d-a95e-11cdd00baf49'}`}
@@ -198,7 +198,7 @@ export const CopySnippet = () => {
                     </span> 
                     <span className="flex-1">
                       <span>
-                        {`</PanoraProviderCard>`}
+                        {`/>`}
                       </span>
                     </span>
                   </span>