+++ packages/api/src/@core/magic-link/magic-link.controller.ts
@@ -32,7 +32,6 @@ export class MagicLinkController {
     return this.magicLinkService.createUniqueLink(data);
   }
 
-  // todo: only admin
   @ApiOperation({
     operationId: 'getMagicLinks',
     summary: 'Retrieve Magic Links',
@@ -43,7 +42,6 @@ export class MagicLinkController {
     return this.magicLinkService.getMagicLinks();
   }
 
-  // admin
   @ApiOperation({
     operationId: 'getMagicLink',
     summary: 'Retrieve a Magic Link',
@@ -52,7 +50,6 @@ export class MagicLinkController {
   @ApiResponse({ status: 200 })
   @Get('single')
   getMagicLink(@Query('id') id: string) {
-    // validate project_id against user
     return this.magicLinkService.getMagicLink(id);
   }
 }