+++ packages/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # api
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [04aec74]
+  - @panora/shared@1.3.0
+
 ## 0.0.7
 
 ### Patch Changes